Renames TM spec
Using fully qualified platform name as module suffix helps with some integration with mobile TM codegen.
