+++ vnext/src-win/src/private/specs/modules/NativePlatformConstantsWindows.js
